+++ client/src/components/Footer/Footer.css
@@ -1,25 +1,34 @@
-footer{
-    color: white;
-    width: 100%;
-    height: 40px;
-    bottom: 0;
-    position: relative;
-    margin-top: auto;
-    justify-content: center;
-    align-content: center;
-    margin: 5px;
-    background-color: black;
-    letter-spacing: 2px;
-    padding: 3px;
+div a {
+  color: white;
+  padding: 2px;
 }
 
-/* div{
-    background-color: black;
-    color: white;
+footer {
+  background-color: black;
+  justify-content: center;
+  align-items: center;
+  display: flex;
+  color: white;
+  padding: 25px;
+}
 
-} */
+footer .github {
+  color: white;
+  width: 30px;
+  padding: 9px;
+  transition: 2s ease-in-out;
+}
 
-div a{
-    color: white;
-    padding: 2px;
+footer .Linkedin {
+  color: white;
+  width: 30px;
+  transition: 2s ease-in-out;
+}
+
+.github:hover {
+  color: purple;
+}
+
+.Linkedin:hover {
+  color: skyblue;
 }
